Fix PDF pattern embedding
The PDF snapshot cow patch was reusing a previously emitted surface
pattern if the surface unique id matched the current surface. This
resulted in incorrect output as the new pattern may have a different
pattern matrix.

This patch fixes the PDF backend to always emit a new pattern but
re-use previously emitted image or metasurface XObjects.