[freelist] Make _cairo_freepool_alloc_from_new_pool static inline.
The xlib boilerplate includes a cairo private header to be able
to disable usage of the render extension.  This indirectly includes
cairo-freelist-private.h which contains a bunch of static inline
functions which use the private _cairo_freepool_alloc_from_new_pool
function, but since that is not inline also, it causes an undefined
extern which cannot be resolved.  The binutils linker doesn't care
since the freelist function aren't actually used in the boilerplate
but the Solaris linker does.  By making the .._alloc_from_new_pool
function inline no dangling references are created and linking
succeeds.